feat(seo): L5.2 SEO baseline (sitemap, robots, OG, JSON-LD, canonical)
- Centralize siteUrl/name/tagline/description in frontend/lib/site.ts
  driven by NEXT_PUBLIC_SITE_URL (defaults to app.thebetterdecision.com).
- Root layout: metadataBase, title template, OG + Twitter defaults,
  robots index/follow.
- Landing: canonical /, summary_large_image Twitter card, and inline
  SoftwareApplication JSON-LD.
- Dynamic OG image at /opengraph-image (1200x630, brand navy + gold).
- Public sitemap.xml and robots.txt listing allowed routes and
  disallowing all authenticated /dashboard, /transactions, /admin, etc.
- Split login + register into server wrapper + client body so each can
  export page-level Metadata with its own canonical.
- Privacy + Terms: title cleaned up for the "%s . The Better Decision"
  template and canonical added.
