+++ .env.example
@@ -58,6 +58,11 @@ GOOGLE_CLIENT_SECRET=
 # Only set this if running frontend standalone without nginx
 NEXT_PUBLIC_API_URL=
 
+# Canonical site URL used for SEO metadata (canonical tags, sitemap, robots,
+# Open Graph image URLs). Production should be the public origin visitors
+# actually reach. Default below matches the current DO App Platform deployment.
+NEXT_PUBLIC_SITE_URL=https://app.thebetterdecision.com
+
 # --- Billing ---
 # DEFAULT_PLAN_SLUG=pro        # "pro" during beta, "free" when billing goes live
 # TRIAL_DURATION_DAYS=14
