+++ frontend/app/page.tsx
@@ -1,19 +1,54 @@
 import type { Metadata } from "next";
 import LandingPageBody from "@/components/landing/LandingPageBody";
+import { siteDescription, siteName, siteTagline, siteUrl } from "@/lib/site";
+
+const pageTitle = `${siteName}: ${siteTagline}`;
 
 export const metadata: Metadata = {
-  title: "The Better Decision: know your money, plan what's next",
-  description:
-    "A finance app for normal people. Know what you have, what's coming, and where it goes. No spreadsheet fatigue.",
+  title: {
+    absolute: pageTitle,
+  },
+  description: siteDescription,
+  alternates: {
+    canonical: "/",
+  },
   openGraph: {
-    title: "The Better Decision: know your money, plan what's next",
-    description:
-      "A finance app for normal people. Know what you have, what's coming, and where it goes.",
-    type: "website",
-    siteName: "The Better Decision",
+    url: "/",
+    title: pageTitle,
+    description: siteDescription,
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: pageTitle,
+    description: siteDescription,
+  },
+};
+
+const jsonLd = {
+  "@context": "https://schema.org",
+  "@type": "SoftwareApplication",
+  name: siteName,
+  description: siteDescription,
+  applicationCategory: "FinanceApplication",
+  operatingSystem: "Web",
+  url: siteUrl,
+  offers: {
+    "@type": "Offer",
+    price: "0",
+    priceCurrency: "EUR",
+    availability: "https://schema.org/InStock",
+    description: "14-day free trial",
   },
 };
 
 export default function LandingPage() {
-  return <LandingPageBody />;
+  return (
+    <>
+      <script
+        type="application/ld+json"
+        dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
+      />
+      <LandingPageBody />
+    </>
+  );
 }
