+++ frontend/lib/site.ts
@@ -1,0 +1,10 @@
+export const siteUrl =
+  process.env.NEXT_PUBLIC_SITE_URL?.replace(/\/$/, "") ??
+  "https://app.thebetterdecision.com";
+
+export const siteName = "The Better Decision";
+
+export const siteTagline = "know your money, plan what's next";
+
+export const siteDescription =
+  "A finance app for normal people. Know what you have, what's coming, and where it goes. No spreadsheet fatigue.";
